move common/ to components/datastructs/
